代码清理，试着更改枪械网络代码，但未经测试

+++ Content/Projectiles/Weapons/Melee/HyperBlade.cs
@@ -1,9 +1,9 @@
 ﻿using CalamityMod;
-using Terraria.ID;
-using Terraria;
-using Terraria.ModLoader;
-using Microsoft.Xna.Framework;
 using CalamityOverhaul.Common;
+using Microsoft.Xna.Framework;
+using Terraria;
+using Terraria.ID;
+using Terraria.ModLoader;
 
 namespace CalamityOverhaul.Content.Projectiles.Weapons.Melee
 {
@@ -41,8 +41,8 @@ namespace CalamityOverhaul.Content.Projectiles.Weapons.Melee
         public override Color? GetAlpha(Color lightColor) {
             if (Projectile.timeLeft < 85) {
                 byte b2 = (byte)(Projectile.timeLeft * 3);
-                byte a2 = (byte)(100f * ((float)b2 / 255f));
-                return new Color((int)b2, (int)b2, (int)b2, (int)a2);
+                byte a2 = (byte)(100f * (b2 / 255f));
+                return new Color(b2, b2, b2, a2);
             }
             return new Color(255, 255, 255, 100);
         }
